+++ examples/HHTut/HHTut.py
@@ -66,6 +66,7 @@ simConfig.duration = 1*1e3 # Duration of the simulation, in ms
 simConfig.dt = 0.025 # Internal integration timestep to use
 simConfig.seeds = {'conn': 1, 'stim': 1, 'loc': 1} # Seeds for randomizers (connectivity, input stimulation and cell locations)
 simConfig.verbose = False  # show detailed messages 
+simConfig.hParams = {'v_init': -75}
 
 # Recording 
 simConfig.recordCells = []  # which cells to record from