+++ netpyne/analysis/spikes.py
@@ -293,8 +293,8 @@ def plotSyncs (include =['allCells', 'eachPop'], timeRanges = None, timeRangeLab
 # -------------------------------------------------------------------------------------------------------------------
 @exception
 def plotRaster (include = ['allCells'], timeRange = None, maxSpikes = 1e8, orderBy = 'gid', orderInverse = False, labels = 'legend', popRates = False,
-        spikeHist = None, spikeHistBin = 5, syncLines = False, lw = 2, marker = '|', markerSize=5, popColors = None, figSize = (10,8), dpi = 100, saveData = None, saveFig = None,
-        showFig = True):
+        spikeHist=None, spikeHistBin=5, syncLines=False, lw=2, marker='|', markerSize=5, popColors=None, figSize=(10, 8), fontSize=12,
+        dpi = 100, saveData = None, saveFig = None, showFig = True):
     '''
     Raster plot of network cells
         - include (['all',|'allCells',|'allNetStims',|,120,|,'E1'|,('L2', 56)|,('L5',[4,5,6])]): Cells to include (default: 'allCells')
@@ -334,10 +334,18 @@ def plotRaster (include = ['allCells'], timeRange = None, maxSpikes = 1e8, order
 
     if isinstance(orderBy, basestring) and orderBy not in cells[0]['tags']:  # if orderBy property doesn't exist or is not numeric, use gid
         orderBy = 'gid'
+    elif orderBy is 'pop':
+        df['popInd'] = df['pop'].astype('category')
+        df['popInd'].cat.set_categories(sim.net.pops.keys(), inplace=True)
+        orderBy='popInd'
     elif isinstance(orderBy, basestring) and not isinstance(cells[0]['tags'][orderBy], Number):
         orderBy = 'gid'
-
+        
     if isinstance(orderBy, list):
+        if 'pop' in orderBy:
+            df['popInd'] = df['pop'].astype('category')
+            df['popInd'].cat.set_categories(sim.net.pops.keys(), inplace=True)
+            orderBy[orderBy.index('pop')] = 'popInd'
         keep = keep + list(set(orderBy) - set(keep))
     elif orderBy not in keep:
         keep.append(orderBy)
@@ -418,15 +426,18 @@ def plotRaster (include = ['allCells'], timeRange = None, maxSpikes = 1e8, order
         timeRange[1] =  sel['spkt'].max()
 
     # Calculate spike histogram
-
     if spikeHist:
         histo = np.histogram(sel['spkt'].tolist(), bins = np.arange(timeRange[0], timeRange[1], spikeHistBin))
         histoT = histo[1][:-1]+spikeHistBin/2
         histoCount = histo[0]
-    # Plot spikes
-    fig,ax1 = plt.subplots(figsize=figSize)
-    fontsiz = 12
 
+    # Plot spikes
+    # set font size
+    plt.rcParams.update({'font.size': fontSize})
+
+    fig,ax1 = plt.subplots(figsize=figSize)
+    fontsiz = fontSize
+ 
     if spikeHist == 'subplot':
         gs = gridspec.GridSpec(2, 1,height_ratios=[2,1])
         ax1=plt.subplot(gs[0])
@@ -460,9 +471,9 @@ def plotRaster (include = ['allCells'], timeRange = None, maxSpikes = 1e8, order
     if syncLines:
         for spkt in sel['spkt'].tolist():
             ax1.plot((spkt, spkt), (0, len(cells)+numNetStims), 'r-', linewidth=0.1)
-        plt.title('cells=%i syns/cell=%0.1f rate=%0.1f Hz sync=%0.2f' % (numCells,connsPerCell,firingRate,syncMeasure()), fontsize=fontsiz)
+        plt.title('cells=%i   syns/cell=%0.1f   rate=%0.1f Hz   sync=%0.2f' % (numCells,connsPerCell,firingRate,syncMeasure()), fontsize=fontsiz)
     else:
-        plt.title('cells=%i syns/cell=%0.1f rate=%0.1f Hz' % (numCells,connsPerCell,firingRate), fontsize=fontsiz)
+        plt.title('cells=%i   syns/cell=%0.1f   rate=%0.1f Hz' % (numCells,connsPerCell,firingRate), fontsize=fontsiz)
     # Axis
     ax1.set_xlabel('Time (ms)', fontsize=fontsiz)
     ax1.set_ylabel(ylabelText, fontsize=fontsiz)
@@ -602,7 +613,10 @@ def plotSpikeHist (include = ['allCells', 'eachPop'], timeRange = None, binSize 
 
     # Plot separate line for each entry in include
     for iplot,subset in enumerate(include):
-        cells, cellGids, netStimLabels = getCellsInclude([subset])
+        if isinstance(subset, list):
+            cells, cellGids, netStimLabels = getCellsInclude(subset)
+        else:
+            cells, cellGids, netStimLabels = getCellsInclude([subset])
         numNetStims = 0
 
         # Select cells to include
@@ -656,7 +670,10 @@ def plotSpikeHist (include = ['allCells', 'eachPop'], timeRange = None, binSize 
 
         histData.append(histoCount)
 
-        color = popColors[subset] if subset in popColors else colorList[iplot%len(colorList)]
+        if isinstance(subset, list): 
+            color = colorList[iplot%len(colorList)]
+        else:   
+            color = popColors[subset] if subset in popColors else colorList[iplot%len(colorList)]
 
         if not overlay:
             plt.subplot(len(include),1,iplot+1)  # if subplot, create new subplot
@@ -683,7 +700,10 @@ def plotSpikeHist (include = ['allCells', 'eachPop'], timeRange = None, binSize 
     # Add legend
     if overlay:
         for i,subset in enumerate(include):
-            color = popColors[subset] if subset in popColors else colorList[i%len(colorList)]
+            if isinstance(subset, list):
+                color = colorList[i%len(colorList)]
+            else:
+                color = popColors[subset] if subset in popColors else colorList[i%len(colorList)]
             plt.plot(0,0,color=color,label=str(subset))
         plt.legend(fontsize=fontsiz, bbox_to_anchor=(1.04, 1), loc=2, borderaxespad=0.)
         maxLabelLen = min(10,max([len(str(l)) for l in include]))
@@ -726,7 +746,7 @@ def plotSpikeHist (include = ['allCells', 'eachPop'], timeRange = None, binSize 
 @exception
 def plotSpikeStats (include = ['allCells', 'eachPop'], statDataIn = {}, timeRange = None, graphType='boxplot', stats = ['rate', 'isicv'], bins = 50,
                  popColors = [], histlogy = False, histlogx = False, histmin = 0.0, density = False, includeRate0=False, legendLabels = None, normfit = False,
-                 fontsize=14, histShading=True, xlim = None, dpi = 100, figSize = (6,8), saveData = None, saveFig = None, showFig = True, **kwargs): 
+                  histShading=True, xlim = None, dpi = 100, figSize = (6,8), fontSize=12, saveData = None, saveFig = None, showFig = True, **kwargs): 
     ''' 
     Plot spike histogram
         - include (['all',|'allCells','allNetStims',|,120,|,'E1'|,('L2', 56)|,('L5',[4,5,6])]): List of data series to include. 
@@ -747,17 +767,16 @@ def plotSpikeStats (include = ['allCells', 'eachPop'], statDataIn = {}, timeRang
     '''
 
     from .. import sim
-
     print('Plotting spike stats...')
 
     # Set plot style
     colors = []
     params = {
-        'axes.labelsize': 14,
-        'font.size': 14,
-        'legend.fontsize': 14,
-        'xtick.labelsize': 14,
-        'ytick.labelsize': 14,
+        'axes.labelsize': fontSize,
+        'font.size': fontSize,
+        'legend.fontsize': fontSize,
+        'xtick.labelsize': fontSize,
+        'ytick.labelsize': fontSize,
         'text.usetex': False,
         }
     plt.rcParams.update(params)
@@ -776,7 +795,7 @@ def plotSpikeStats (include = ['allCells', 'eachPop'], statDataIn = {}, timeRang
     for stat in stats:
         # create fig
         fig,ax1 = plt.subplots(figsize=figSize)
-        fontsiz = fontsize 
+        fontsiz = fontSize
         xlabel = xlabels[stat]
 
         statData = []
@@ -846,8 +865,7 @@ def plotSpikeStats (include = ['allCells', 'eachPop'], statDataIn = {}, timeRang
                     else:
                         rates = [spkinds.count(gid)*toRate for gid in set(spkinds)] \
                             if len(spkinds)>0 else [0] #cellGids] #set(spkinds)] 
-
-                    statData.insert(0, rates)
+                    statData.append(rates)
 
 
                 # Inter-spike interval (ISI) coefficient of variation (CV) stats
@@ -857,7 +875,7 @@ def plotSpikeStats (include = ['allCells', 'eachPop'], statDataIn = {}, timeRang
                         for gid in set(spkinds)]
                     isimat = [[t - s for s, t in zip(spks, spks[1:])] for spks in spkmat if len(spks)>10]
                     isicv = [np.std(x) / np.mean(x) if len(x)>0 else 0 for x in isimat] # if len(x)>0] 
-                    statData.insert(0, isicv) 
+                    statData.append(isicv) 
 
 
                 # synchrony
@@ -880,7 +898,7 @@ def plotSpikeStats (include = ['allCells', 'eachPop'], statDataIn = {}, timeRang
                         syncMat = np.mean(pyspike.spike_sync_matrix(spkmat), 0)
                         
 
-                    statData.insert(0, syncMat)
+                    statData.append(syncMat)
 
             colors.insert(0, popColors[subset] if subset in popColors 
                 else colorList[iplot%len(colorList)])  # colors in inverse order
@@ -895,7 +913,7 @@ def plotSpikeStats (include = ['allCells', 'eachPop'], statDataIn = {}, timeRang
         if graphType == 'boxplot':
             meanpointprops = dict(marker=(5,1,0), markeredgecolor='black', markerfacecolor='white')
             labels = legendLabels if legendLabels else include
-            bp=plt.boxplot(statData, labels=labels, notch=False, sym='k+', meanprops=meanpointprops, 
+            bp=plt.boxplot(statData, labels=labels[::-1], notch=False, sym='k+', meanprops=meanpointprops, 
                         whis=1.5, widths=0.6, vert=False, showmeans=True, patch_artist=True)
             plt.xlabel(xlabel, fontsize=fontsiz)
             plt.ylabel('Population', fontsize=fontsiz) 
@@ -1079,9 +1097,9 @@ def plotSpikeStats (include = ['allCells', 'eachPop'], statDataIn = {}, timeRang
 # -------------------------------------------------------------------------------------------------------------------
 ## Plot spike histogram
 # -------------------------------------------------------------------------------------------------------------------
-exception
-def plotRatePSD (include = ['allCells', 'eachPop'], timeRange = None, binSize = 5, maxFreq = 100, NFFT = 256, noverlap = 128, smooth = 0, overlay=True, ylim = None, 
-    popColors = {}, figSize = (10,8), saveData = None, saveFig = None, showFig = True): 
+@exception
+def plotRatePSD(include=['allCells', 'eachPop'], timeRange=None, binSize=5, maxFreq=100, NFFT=256, noverlap=128, smooth=0, overlay=True,
+    ylim = None, popColors = {}, fontSize=12, figSize=(10,8), saveData=None, saveFig=None, showFig=True): 
     ''' 
     Plot firing rate power spectral density (PSD)
         - include (['all',|'allCells','allNetStims',|,120,|,'E1'|,('L2', 56)|,('L5',[4,5,6])]): List of data series to include. 
@@ -1122,9 +1140,13 @@ def plotRatePSD (include = ['allCells', 'eachPop'], timeRange = None, binSize = 
 
     # create fig
     fig,ax1 = plt.subplots(figsize=figSize)
-    fontsiz = 12
-    
-    allPower, allSignal, allFreqs=[], [], []
+    fontsiz = fontSize
+
+    # set font size
+    plt.rcParams.update({'font.size': fontSize})
+        
+    allPower, allSignal, allFreqs = [], [], []
+
     # Plot separate line for each entry in include
     for iplot,subset in enumerate(include):
         cells, cellGids, netStimLabels = getCellsInclude([subset])