+++ CHANGES.md
@@ -1,4 +1,14 @@
+# Version 0.9.3
+
+- Improved conversion from SONATA format, including 300 biophys cell example
+
+- Added saveCellConns, pt3dRelativeToCellLocation and invertedYCoord cfg options
+
+- Fixed bug plotting traces with one fig per trace
+
 # Version 0.9.2
+
+- Support for saving simulation at intervals  
 
 - Improved Python cell template import so can read section names from a Python list or dict of sections
 
@@ -8,6 +18,38 @@
 
 - Added option cfg.distributeSynsUniformly to locate synapses at uniformly across section list; if false, place one syn per section in section list 
 
+- Updated RxD calcium buffering example
+
+- Added netClamp example to re-simulate activity of single cell from network data
+
+- Added support for recording and plotting variable time step
+
+- Replaced spectrogram FFT method with Morlet wavelet method
+
+- Updated installation instructions
+
+- Added 'fontSize' argument to several plots
+
+- Fixed issue completing batches in Python 3
+
+- Fixed bug saving evolutionary optimization batch output file
+
+- Fixed bug in plotShape()
+
+- Fixed bug in plotSpikeStats()
+
+- Fixed bug loading files with colons
+
+- Fixed issue sorting raster by population and y coord
+
+
+# Version 0.9.1.3
+
+- Removed deprecated hold function from plotConn
+
+# Version 0.9.1.2
+
+- Set 'pandas==0.23.4' to avoid error 
 
 # Version 0.9.1.1